debug shenanigans
added some  images to create a visual grid. showing the exact cutoff points for all of the grid pieces. these are also part of the grid, and will be used to create the coordinates to be used for the game.

for now, also made the pieces snap to the grid. visually looks like the pieces snap into the squares created by the visual aid.

Bug: will be irrelevant later but when a piece is placed on a negative coordinate (either x or y, doesn't have to be both) it will snap to the next closest x/y coordinate towards the positive positions.