Wrong drop bug fix
When dropping pieces, sometimes, the held piece would snap to the top left corner (0,0) instead of where you held it. This was because it took the 'oldPosition' of the piece it was hovering over. However, untouched pieces haven't set that variable for themselves yet, so it would be equal to 0. This was fixed by taking that pieces current global position instead.