Debugging and matching prep
All pieces now have their internal grid pos variable updated alongside the 2D array whenever a move is made. This is a decoupled function that can be called for any single piece.
Additionally, to check if all pieces are correctly stored, I created a function to go through the 2D array and print out the contents to represent the grid visually in the debug console. Also works with unevenly shaped boards like 10x4 or 4x5