Added gravity reaction and basic refill
Now, when a match is made, pieces fall down into empty spots, leaving the gaps at the top.
These gaps are then filled in by the new refill function. To make this function, I repurposed part of the existing generation function to create a function that simply creates a single new piece. This piece making function is now used in both the initial generation, as well as the new refill function use this to create new pieces.