Added "refill algorithms"
This is what I'm talking about

Added a new setting "refill algorithms".
With this setting, you can now customize the way new pieces are generated.
By default, piece type is selected randomly. The only other valid algorithm currently is 'order', refilling the board in a predetermined order. It's a bit nonsensical gameplay wise, but it does work as it says on the tin.