Added 'balanced' algorithm
This algorithm counts how many of each type are on the board and increases the chances for it to appear proportionally to its rarity on the board. So pieces of which there are less are more likely to appear and pieces of which there are already a lot of appear more often.

Also, small collision rework. The dragged tile now only looks at the tile it touches. This way. you can't accidentally overlap with multiple pieces, which infinitely try to become the 'swapped piece'