Added assisting and fighting refill algorithms
- Added a function to check for a valid match in a specific position. On its own, not very useful. But when combined with shenanigans, can identify computer cascades and possible moves.
- Added a function that uses the previous to check for all available moves on the board. It gives those moves in a list with a from and to position.
- Added a variant of the move checker that checks all available moves from a specific spots, and returns whether or not there is a move available there.
- Added an 'assisting' and 'fighting' refill algorithm.
They both check the spots to be refilled for available moves, given a piece type, and then make that type available as an option for the refill or not depending on the result. The assist will only pick pieces that do allow for a new match to be made, and the fighter will only pick pieces that don't.